Organize all test files into a separet test folder
