+++ src/tests/test-pages/Home.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import Home from '../pages/Home';
+import Home from './../../components/pages/Home';
 
 it('render home component', () => {
   const home = renderer.create(<Home />).toJSON();
