+++ src/tests/test-pages/Quote.test.js
@@ -1,5 +1,5 @@
 import renderer from 'react-test-renderer';
-import Quote from '../pages/Quote';
+import Quote from '../../components/pages/Quote';
 
 it('should render Quotes component', () => {
   const quote = renderer.create(<Quote />).toJSON();
